Reduce occurences of cancelled context check in database
In resolveLocked(), we used to check whether the context was
cancelled (with task.StopReason()) in every resolve. This change makes
resolveLocked() only return a cancelled context error when the
first-time resolve (the "build" one) is actually cancelled, as checked
by task.ShouldStop().

This is an optimization: by avoiding to call task.StopReason() in
every database resolve, we speed up these resolve.