Disable the UI error pop-up window for a known SWT issue.
There's a known SWT's compatibility issue with macOS Big Sur, a more
detailed context could be found in the bug ticket. Based on testing,
this NPE UI error triggered by GlCanvas object is not that harmful: it
will not break gapis and hang the AGI session, nor will the rendering
get affected. But the error dialog which keeps popping up is a bit annoying.
This commit will suppress the dialog popping for this bug, while the error
stack trace will still get printed into terminal and log file silently.

Bug: b/178397207.