Implement GPU counter data's detail visualization in perftab.
Add a canvas in the perftab to draw GPU counter detail curve. The curve
correspond to the selected cell in the performance table.
This detail counter curve surves mainly for 2 purposes:
1. Help improve those non-time based rate counter's aggregation.
With the detailed graph, users could notice a counter's spike inside a
long renderpass, otherwise the spike may gets flattened and loses attention
with merely the performance number.
2. Help AGI developers or early perftab users notice whether there's
calculation bugs on the performance number.

Bug: b/158057709.
Test: Manual test by openning the Performance(Experimental) tab.