Remove default cases in sType selects.
This has the effect of no longer simply ignoring unknown sTypes in the
pNext handling. Before this, any unknown sType struct would be ignored.
Without the default case, mutate will now fail on unknown sTypes. This
is preferable, since we were likely not observer memory properly, when
ignoring, which could lead to odd/crashing behavior in the replayer.