Fix an AGI freezing and OOM problem when loading block memory.
This bug may happen when AGI has Memory tab checked, but not put at front.

When block memory UI tries to repaint, it will 1.check repainting area, 2.
calculate repainting memory range accordingly, 3.query the server to load the
memory range.
In AGI, we have been using gc.getClipping() to do the step #1 for quite some
time, but recently we find that it will not do #1 correctly when the Memory tab
is hidden in the back, and thus the following steps will get wrong and cause
incorrect server query. (It's possible but not sure, that SWT handles gc.getClipping()
differently in newer version and thus this bug surfaces recently.) This PR tries
to do the step #1 based on another data source and thus solve the problem.

Test: This bug may happen when AGI has Memory tab checked, but not put at
front. Reproduced on MacOS BigSur.
Bug: b/181012589.