Remove gapil extern function hasDynamicProperty()
This extern was used to check if a given dynamic state was enabled or
not by iterating over the relevant argument to
vkCreateGraphicsPipelines().

However, we do record which dynamic state is enabled as part of the
pipeline state anyway, so there is no need for an extern function
here. This change simply makes sure to record the enabled dynamic
state early on, and then check which one are enabled by reusing
hasDynamicState(). This lets us get rid of hasDynamicProperty().