Add trace option: name of the process to capture
Some games may start with a first process (or Android activity), and
then fork another process/activity where the actual game rendering
happens. If the starting process also uses Vulkan, then it is likely
to be the first to establish a gapii-gapis connection, such that AGI
won't be able to capture the actual game.

This change adds a "process name" trace option to indicate the name of
the process to capture. This is a new text entry in the UI's
TracerDialog. The value gets propagated as a trace option through
gapis, and is eventually passed to gapii: via an Android system
property on Android, or via an environment variable (refactoring the
existing `GAPID_CAPTURE_PROCESS_NAME` mechanism) on Desktop. The gapii
Spy, upon creation, looks for this process name and only establish a
connection with gapis if the current process name matches the one to
capture, or if the capture process name is empty.

This was motivated by a Unity game which forks a new process whose
name is `com.example.game:UnityKillsMe`, where the game is actually
rendered. This seems to be a way to include Unity games into an
Android app, the `UnityKillsMe` suffix lets Unity take care of killing
the forked activity upon app shutdown.

Bug: b/177363287
Test: manual