[GLES Cleanup] Remove gapii spy init on JNI_OnLoad
GAPID used to create the Spy upon JNI_OnLoad(), which was called when
libgapii was loaded from Java when GLES interception was initialized
using the Java debugger.

Since AGI only deals with Vulkan, this code is now irrelevant. I
double checked that it is never called when we do Vulkan capture, and
that our capture still works once this code is removed.

Bug: n/a
Test: manual