Refactors the ETC proto structs to simplify the code around ETC format image creation
This is a work continuation of #773

The way that ETC protos structured, end up with lots of constructors that are not really necessary and to avoid duplicate them, actual converters and the image format definitions were in different packages that caused b/184954321

While investigating that bug, I have realized the situation and created an initial fix. This requires more work later that I discuss in the design doc attached to refactor work's bug tracker.

For a summary, all ETC formats are under one struct and the information that buried into their name now lies as fields in proto, which ended up having much more simpler code to interact with it.

Refactor Work Bug Tracking: b/186110226