Also set the mip level to zero when reading from the staging image.
When we need to do a blit to read back the framebuffer, the staging
image will have only one layer and one level. We correctly reset
the layer for the copy to 0, but didn't set the level to 0. Thus
we would attempt to copy from an invalid mip level if we blitted
from anything but level 0.