Fix a bug when setting the end of the framebuffer histogram's default range.
AGI will calculate and set default histogram range in framebuffer view and the general
intention is to set the range's end to the bin at 99% percentile in the histogram.

However, different channels will reach 99% percentile at different bin, and previously,
because of an early return, when iterating through the bins, if AGI found any of the
channels already reaches 99% percentile, it will immediately return that bin as the end
of the whole range. This will cause trouble in cases where different channels reaches
99% at very different bins. e.g. for a R16G16 image, the blue channel is always 0,
so 100% of the pixels falls at the value 0, and thus blue channel's 99% percentile bin reaches
early at bin 0, if we use this as the default histogram range's end, it will erase detail
for the red and yellow channel.

Ideally the range's end should be the last 99% percentile bin across all channels, so
that the main distribution of all channels will get displayed.This PR makes sure the range
starts at the *first* 1% percentile bin, and ends at the *last* 99% percentile bin.

Bug: b/177572847.