Avoid the need to restart AGI after installing ANGLE
Before this change, if you start AGI with a device that has no ANGLE,
then use AGI to install ANGLE, then try and start an ANGLE trace,
you'll get a "ANGLE requested but no ANGLE package found" error. This
is because installing ANGLE doesn't update the device info, such that
AGI still thinks ANGLE is not present on the device.

The way device info is stored is non-trivial: there's a background
task (see monitorAndroidDevices()) that monitors devices and store
them in a "registry" that is queried when we list devices. It sounds
like the proper way to update device info is to remove the device from
the registry, and let the monitor re-scan it accordingly. It sounds
relevant to do it after any APK install that may change the device
info.

On the UI side, we also need to make sure to refresh the device info
after installing ANGLE.

Bug: b/186611374
Test: manual, making sure to run AGI on a device with no ANGLE on it.