Create group for draw cmds without state setup
For consecutive vkCmdDraw* commands, create a draw group in the command
tree for the second draw (without any state setup) so it matches how
other draws (with state setup) are displayed. There will only be a
single draw command in the group for this case.

Also, create draw groups for vkCmdClearAttachements since it behaves
much like a draw.

Fixes #800