Workaround for a race condition with perfetto slices on arm mali
There is a problem with the AGI/Perfetto integration for ARM where the
mali driver does not appear to respect the perfetto producer ready
signal. Previously a sleep was added to the code to work around this
race condition. This CL enlarges the wait time for devices that appear
to have slower drivers and adds a second wait at termination because I
am starting to suspect there is a dual of the first race condition on
exiting, where we may be processing the data before it has all been read
back. This is a temporary workaround until we get something to address
the real issue.

Bugs:
b/188440357
b/190170108