Worksaround the mapping bug due to duplicated Vulkan Handles.
This does not solve the duplication but minimises the risk of
duplication by not mapping the initial command handles for VkFramebuffer.

In Mapping Handler, there is no distinction between initialCmds and
Trace Commands, therefore lots of handles that are being created and
destroyed during the initialCmds also being mapped.

This PR, checks the VkFramebuffer handles and only maps them if they
are used by the application itself.

b/180532468