Add state rebuilder part for VK_EXT_line_rasterization
Add missing part for support of VK_EXT_line_rasterization: make sure
to request this extension's features when creating devices as part of
the state rebuilder.

Also need to declare pNext as `void*` (and not `const void*`) at the
GAPIL level such that the Golang code passes type checking. This is
consistent with other vulkan structs in GAPIL that always use `void*`
as a type for their `pNext` field.

Bug: b/192662962