Fix VK_EXT_line_rasterization support
This change makes sure to add the relevant VK_EXT_line_rasterization
info when re-creating graphics pipelines in the state rebuilder. Also,
the info related to this extension is now stored in a separate
PipelineRasterizationLineStateCreateInfoEXT class, which is used as a
member in the graphics pipeline's RasterizationState (whereas before,
all VK_EXT_line_rasterization-related filed were direct members of
RasterizationState).

Bug: b/194180484