Document how the framegraph is computed
This documentation is done in the function comment of GetFramegraph(),
which is the API interface method that acts as an entry point to the
whole framegraph creation process. When developers browse the code,
GetFramegraph() is most likely the place where they will end up when
looking up how the framegraph is computed.

Also, minor comment updates.

Bug: n/a