Look for queues on the same device as the command buffer.
When handing the frame buffer read back, we need to find a queue to
execute work on. If we are limited by a provided command buffer, we look
for a queue of the same family as the command buffer's pool. However,
we incorrectly considered queues from *all* devices when looking for a
queue with the same family, rather than only looking at queues with the
same family from the same device as the command buffer.