Use the common shader widget in the pipeline view.
Since the shader view can show the static analysis, this also
includes a change on the server side to pack the static analysis
data with the shader, rather than in its own DataGroup.

Also makes the formatting in the pipeline view allow things to size
themselves based on thier size, rather than all equal. This allows
the shader code to take up more space, which is currently wasted
as white space.