Rework texture resource UX.
Show the texture image in the texture view as we used to, but allow
textures to be pinned into their own tabs. This gets rid of the blank
"Texture" tab, but still allows viewing textures in their own tabs.
Pinning is done via a new button in the button bar in the image UI.