Modify protected feature and queue queries
Protected memory is not currently supported by capture/replay and it is
currently unknown if it can be supported. In the meantime, use the spy
layer to suppress support for protected memory advertised by the
underlying driver. This has to be done by overriding
vkGetPhysicalDeviceFeatures2 and
vkGetPhysicalDeviceQueueFamilyProperties since protected memory is a
core Vulkan 1.1 feature and not an extension.