Refactor command splitter
This refactor rewrites the command splitter extensively
without changing the general logic or observable behaviour.

It mostly aims about increasing readability with more explanation via
longer variable names(It's not very Gopher way, yes), additional comments
to explain what happens and why it happens for the people new to Vulkan and
bringing it to the coding style(which is different than Go)
after we had previous general transform refactor.

There are some changes in when we actually rewrite the renderpasses especially
with secondary command buffers. The general principle is do not rewrite the
renderpasses if there is no command requested in it to keep replay
more faithfully to the original