Ensure resource IDs for MEC resources are unique.
If the first command in the trace creates resources, their IDs would match
the IDs generated for the resources serialized as part of the MEC state.
This happened because the command ID used to generate the resource IDs
for the MEC resources was the same as that of the first command. This would
result in the first MEC resource to get ID 0xdd3d... and the first
resource of the traced commands to get a matching ID. This change fixes
this, by ensuring the MEC resources get gen'ed off a unique command ID.