Fixes an issue with MEC resources' creation command reference.
We track at what command resources are created. Resources created before
the trace (in the MEC state) do not have a "creation command". In the past
they were associated with the first command in the trace, which has caused
some minor reporting type problems, but also later on a crash. This was
partially fixed in 21f0184e9, but together with e313301e3 caused a
regression in the pipeline view, which now was no longer properly referring
to shader resources due to the now negative creation reference.