Only create the required encoding functions.
This changes the encoder plugin to only create encoding functions for
types that actually need them. It does this by recusively walking the
dependencies of the state object, the command and command call
objects, and any command extras and only creates encode functions
for types encountered during the walk. These types represent the set of
all types serialized by the spy.

This significanlty reduces the size of the proto descriptor buffer stored
in the output, as well as the code size of the generated file.